+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 - [added]: Base theming behavior 
 - [added]: Base styling behavior
 - [added]: Types from react-native
+- [added]: Apply to function
+- [added]: Value for function